+++ algorithms/passive/rpni/src/test/java/de/learnlib/algorithm/rpni/it/MdlDfaIT.java
@@ -20,9 +20,9 @@ import java.util.stream.Collectors;
 
 import de.learnlib.algorithm.rpni.BlueFringeMDLDFA;
 import de.learnlib.query.DefaultQuery;
-import de.learnlib.testsupport.it.learner.AbstractDFAPassiveLearnerIT;
-import de.learnlib.testsupport.it.learner.LearnerITUtil;
-import de.learnlib.testsupport.it.learner.PassiveLearnerVariantList;
+import de.learnlib.testsupport.it.AbstractDFAPassiveLearnerIT;
+import de.learnlib.testsupport.it.util.LearnerITUtil;
+import de.learnlib.testsupport.it.variant.PassiveLearnerVariantList;
 import net.automatalib.alphabet.Alphabet;
 import net.automatalib.automaton.fsa.DFA;