+++ oracles/parallelism/src/test/java/de/learnlib/oracle/parallelism/StaticParallelAdaptiveSULTest.java
@@ -1,0 +1,33 @@
+/* Copyright (C) 2013-2025 TU Dortmund University
+ * This file is part of LearnLib <https://learnlib.de>.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.learnlib.oracle.parallelism;
+
+import java.util.concurrent.atomic.AtomicInteger;
+
+import de.learnlib.oracle.parallelism.Utils.TestSULOutput;
+
+public class StaticParallelAdaptiveSULTest extends AbstractStaticParallelAdaptiveOracleTest<TestSULOutput> {
+
+    @Override
+    protected StaticParallelAdaptiveOracleBuilder<Integer, TestSULOutput> getBuilder() {
+        return ParallelOracleBuilders.newStaticParallelAdaptiveOracle(new TestSUL(new AtomicInteger(-1)));
+    }
+
+    @Override
+    protected TestSULOutput extractTestOutput(TestSULOutput output) {
+        return output;
+    }
+}